Add several repository
add several repository and cleaning code
