+++ src/main/java/com/management_asset/api/model/Asset.java
@@ -1,0 +1,35 @@
+package com.management_asset.api.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "tb_m_asset")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Asset {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+    private String description;
+    private String path_photo_asset;
+
+    @ManyToOne
+    @JoinColumn(name = "category_id", referencedColumnName = "id")
+    private Category category;
+
+    @ManyToOne
+    @JoinColumn(name =  "asset_status_id",  referencedColumnName = "id")
+    private AssetStatus assetStatus;
+}
