+++ src/main/java/com/management_asset/api/model/AssetCondition.java
@@ -1,14 +1,15 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import java.sql.Date;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-
-import org.hibernate.annotations.ManyToAny;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -20,6 +21,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AssetCondition {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
     @OneToMany
     @JoinColumn(name = "asset_id", referencedColumnName = "id")
     private Asset asset;
@@ -35,5 +40,4 @@ public class AssetCondition {
     private Date checking_date;
     private String proof_of_damage;
     private String notes;
-
 }
