+++ src/main/java/com/management_asset/api/model/AssetStatus.java
@@ -1,4 +1,4 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -11,13 +11,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "tb_m_parts")
+@Table(name = "tb_m_asset_status")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Parts {
+public class AssetStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    private String name;
+    private String status;
 }
