+++ src/main/java/com/management_asset/api/model/Category.java
@@ -1,4 +1,4 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -11,14 +11,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "tb_m_loan_status_process")
-@Data 
+@Table(name = "tb_m_category")
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-
-public class LoanStatusProcess {
+public class Category {
     @Id
-    @GeneratedValue (strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    private String loaningStatus;
+    private String category;
 }
