+++ src/main/java/com/management_asset/api/model/Employee.java
@@ -1,4 +1,4 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -13,23 +13,19 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "tb_m_asset")
+@Table(name = "tb_m_employee")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Asset {
+public class Employee {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private String description;
-    private String path_photo_asset;
+    private String email;
 
     @ManyToOne
-    @JoinColumn(name = "category_id", referencedColumnName = "id")
-    private Category category;
-
-    @ManyToOne
-    @JoinColumn(name =  "asset_status_id",  referencedColumnName = "id")
-    private AssetStatus assetStatus;
+    @JoinColumn(name = "manager_id", referencedColumnName = "id")
+    private Employee employee;
 }
