+++ src/main/java/com/management_asset/api/model/LoanStatusHistory.java
@@ -1,4 +1,4 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import java.time.LocalDateTime;
 
@@ -20,9 +20,9 @@ import lombok.NoArgsConstructor;
 public class LoanStatusHistory {
     private LocalDateTime createdDate;
 
-    // @OneToMany
-    // @JoinColumn(name = "approver_id", referencedColumnName = "id")
-    // private Employee approver;
+    @OneToMany
+    @JoinColumn(name = "approver_id", referencedColumnName = "id")
+    private Employee approver;
 
     @ManyToOne
     @JoinColumn(name = "loaning_id", referencedColumnName = "id")
@@ -31,5 +31,4 @@ public class LoanStatusHistory {
     @ManyToOne
     @JoinColumn(name = "loan_status_process_id", referencedColumnName = "id")
     private LoanStatusProcess loanStatusProcess;
-
 }
