+++ src/main/java/com/management_asset/api/model/LoanStatusProcess.java
@@ -1,4 +1,4 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -11,13 +11,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "tb_m_asset_status")
-@Data
+@Table(name = "tb_m_loan_status_process")
+@Data 
 @AllArgsConstructor
 @NoArgsConstructor
-public class AssetStatus {
+
+public class LoanStatusProcess {
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue (strategy = GenerationType.IDENTITY)
     private Integer id;
-    private String status;
+    private String loaningStatus;
 }
