+++ src/main/java/com/management_asset/api/model/Parts.java
@@ -1,13 +1,9 @@
-package com.management_asset.model;
+package com.management_asset.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.MapsId;
-import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -15,20 +11,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "tb_m_employee")
+@Table(name = "tb_m_parts")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Employee {
-
+public class Parts {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private String email;
-
-    @ManyToOne
-    @JoinColumn(name = "manager_id", referencedColumnName = "id")
-    private Employee employee;
-
 }
