+++ src/main/java/com/management_asset/api/repository/PartsRepository.java
@@ -1,0 +1,11 @@
+package com.management_asset.api.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.management_asset.api.model.Parts;
+
+@Repository
+public interface PartsRepository extends JpaRepository<Parts, Integer> {
+    
+}
